change to allow compilation of previous bugfix
git-svn-id: https://wwmm.ch.cam.ac.uk/svn2/groups/alavi/NECI/trunk@17003 3c9ecaa4-0e0d-0410-a8e4-acb18276dcea

+++ vasp_interface.F90
@@ -34,7 +34,7 @@ contains
       do i=1,nKP
          write (6,*) vasp_kpnts(:,i)
       end do 
-      nEl=nint(vasp_nEl*nKP)
+      nEl=vasp_nEl*nKP
 
       allocate(KPNTInd(nStates))
       allocate(eigv(nStates))
